design pattern facade implementado
